Cleanup in `mcap_vendor` package
- Use `${PROJECT_NAME}` instead of confusing `mcap` for library name and
exporting target, to be consistent with all other packages in ROS2.

- Add export for path to the `mcap` includes and export for mcap_vendor
library. It was an issue in downstream packages that `mcap/mcap.hpp` was
not visible since we were not exporting path to it explicitly.

- Removed test section with linters for `mcap_vendor` package. Since
package itself doesn't contain any tests and own source code and we
shouldn't run linters on any third party code. Prevent CI failures.

Signed-off-by: Michael Orlov <michael.orlov@apex.ai>